requirements.txt
cobble==0.1.3
coverage==5.5
Django==2.2
flake8==3.8.4
importlib-metadata==3.4.0
isort==5.7.0
mammoth==1.4.15
mccabe==0.6.1
Pillow==8.1.0
pycodestyle==2.6.0
pyflakes==2.2.0
python-decouple==3.4
python-dotenv==0.15.0
pytils==0.3
pytz==2021.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.0
